# apps/api/internal/handlers/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orpheus API",
    "version": "0.2.0",
...
    "contact": {
      "name": "Orpheus Engineering",
      "url": "https://docs.orpheus.dev",
      "email": "api@orpheus.dev"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://orpheus.dev/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local dev"
    },
    {
      "url": "https://api.orpheus.dev",
      "description": "Production (placeholder)"
    }
  ],
...
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/v1/erasure-requests": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Request GDPR erasure",
        "operationId": "createErasureRequest",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "confirm"
                ],
                "properties": {
                  "scope": {
                    "type": "string",
                    "enum": [
                      "artifact",
                      "job"
                    ]
                  },
                  "artifact_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "job_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "subject_ref": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "confirm": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Scheduled."
          },
          "400": {
            "description": "Invalid / confirm required."
          },
          "404": {
            "description": "Target not found."
          },
          "409": {
            "description": "Legal hold."
          }
        }
      },
      "get": {
        "tags": [
          "system"
        ],
        "summary": "List erasure requests",
        "operationId": "listErasureRequests",
        "responses": {
          "200": {
            "description": "Requests."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Keycloak-issued JWT. Obtain via your org's Keycloak realm's standard OAuth2/OIDC flow and pass in the `Authorization` header as `Bearer <token>`."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Prefix-scoped API key (e.g. `ak_live_xxx`). Issued via `POST /v1/api-keys`; the full secret is returned exactly once at creation."
      }
    },
...
  }
}